+++ src/bp/core/app/api.ts
@@ -15,9 +15,9 @@ import { EventEngine, EventRepository, Event } from 'core/events'
 import { KeyValueStore } from 'core/kvs'
 import { LoggerProvider } from 'core/logger'
 import * as logEnums from 'core/logger/enums'
-import { MappingRepository } from 'core/mapping/mapping-repository'
 import { MediaServiceProvider } from 'core/media'
-import { ConversationService, MessageService } from 'core/messaging'
+import { ConversationService } from 'core/messaging'
+import { ChatService } from 'core/messaging/chat-service'
 import { ModuleLoader } from 'core/modules'
 import { RealtimeService, RealTimePayload } from 'core/realtime'
 import { getMessageSignature } from 'core/security'
@@ -195,14 +195,14 @@ const distributed = (jobService: JobService): typeof sdk.distributed => {
 const experimental = (
   hookService: HookService,
   conversationService: ConversationService,
-  messageService: MessageService,
+  chatService: ChatService,
   renderService: RenderService
 ): typeof sdk.experimental => {
   return {
     disableHook: hookService.disableHook.bind(hookService),
     enableHook: hookService.enableHook.bind(hookService),
     conversations: conversations(conversationService),
-    messages: messages(messageService),
+    messages: messages(chatService),
     render: render(renderService)
   }
 }
@@ -213,9 +213,9 @@ const conversations = (conversationService: ConversationService): typeof sdk.exp
   }
 }
 
-const messages = (messageService: MessageService): typeof sdk.experimental.messages => {
+const messages = (chatService: ChatService): typeof sdk.experimental.messages => {
   return {
-    forBot: messageService.forBot.bind(messageService)
+    forBot: chatService.forBot.bind(chatService)
   }
 }
 
@@ -286,9 +286,8 @@ export class BotpressAPIProvider {
     @inject(TYPES.JobService) jobService: JobService,
     @inject(TYPES.StateManager) stateManager: StateManager,
     @inject(TYPES.ConversationService) conversationService: ConversationService,
-    @inject(TYPES.MessageService) messageService: MessageService,
-    @inject(TYPES.RenderService) renderService: RenderService,
-    @inject(TYPES.MappingRepository) mappingRepo: MappingRepository
+    @inject(TYPES.ChatService) chatService: ChatService,
+    @inject(TYPES.RenderService) renderService: RenderService
   ) {
     this.http = http(httpServer)
     this.events = event(eventEngine, eventRepo)
@@ -301,7 +300,7 @@ export class BotpressAPIProvider {
     this.bots = bots(botService)
     this.ghost = ghost(ghostService)
     this.cms = cms(cmsService, mediaServiceProvider)
-    this.experimental = experimental(hookService, conversationService, messageService, renderService)
+    this.experimental = experimental(hookService, conversationService, chatService, renderService)
     this.security = security()
     this.workspaces = workspaces(workspaceService)
     this.distributed = distributed(jobService)
