+++ src/bp/core/types.ts
@@ -90,6 +90,7 @@ const TYPES = {
   MessageRepository: Symbol.for('MessageRepository'),
   ConversationRepository: Symbol.for('ConversationRepository'),
   MessageService: Symbol.for('MessageService'),
+  ChatService: Symbol.for('ChatService'),
   ConversationService: Symbol.for('ConversationService'),
   RenderService: Symbol.for('RenderService'),
   MappingRepository: Symbol.for('MappingRepository')
